+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [0.10.0] - 2025-03-21
+
+### Added
+
+- `ProcessEvent()` method to `StateMachine<TContext, TEvent, TState>` and `DisposableStateMachine<TContext, TEvent, TState>`
+- `ProcessEventAsync()` method to `AsyncStateMachine<TContext, TEvent, TState>`
+- `ProcessingResult` enum to represent the outcome of processing an event (NotProcessed, Remained, Transitioned)
+
 ## [0.9.0] - 2023-10-05
 
 ### Added
@@ -86,7 +94,9 @@
 
 - Basic building blocks for generic state machines: `StateMachine<…>`, `IState<…>`, `IPolicy<…>`.
 
-[Unreleased]: https://github.com/qbit86/machinery/compare/machinery-0.9.0...HEAD
+[Unreleased]: https://github.com/qbit86/machinery/compare/machinery-0.10.0...HEAD
+
+[0.10.0]: https://github.com/qbit86/machinery/compare/machinery-0.9.0...machinery-0.10.0
 
 [0.9.0]: https://github.com/qbit86/machinery/compare/machinery-0.8.0...machinery-0.9.0